# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "WeConnect provides a platform that brings businesses and individuals together.",
    "version": "1.0.0",
    "title": "WeConnect",
    "contact": {
      "email": "longe.pelumi@yahoo.com"
    }
  },
  "host": "localhost:8000",
  "basePath": "/api/v1",
  "tags": [
    {
      "name": "User",
      "description": "Signup and Signin Users"
    },
    {
      "name": "Business",
      "description": "Access to Businesses"
    },
    {
      "name": "Review",
      "description": "Access to reviews"
    }
  ],
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
    "/auth/signup": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Registers a user",
        "description": "This endpoint uses a post request to register a new user",
        "operationId": "signup",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Create user object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/userSignup"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User successfully created",
            "schema": {
              "$ref": "#/definitions/signupResponse"
            }
          },
          "400": {
            "description": "Invalid email address",
            "schema": {
              "$ref": "#/definitions/wrongEmail"
            }
          },
          "406": {
            "description": "Invalid Input",
            "schema": {
              "$ref": "#/definitions/invalidInput"
            }
          },
          "409": {
            "description": "Existing email",
            "schema": {
              "$ref": "#/definitions/existingEmail"
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Logs in a user",
        "description": "This endpoint logs in a user",
        "operationId": "login",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Login user object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/userLogin"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Log in user",
            "schema": {
              "$ref": "#/definitions/login"
            }
          },
          "401": {
            "description": "Wrong password",
            "schema": {
              "$ref": "#/definitions/wrongPassword"
            }
          },
          "404": {
            "description": "User not found",
            "schema": {
              "$ref": "#/definitions/UserNotFound"
            }
          },
          "406": {
            "description": "Invalid Input",
            "schema": {
              "$ref": "#/definitions/invalidInput"
            }
          }
        }
      }
    },
    "/auth/user": {
      "put": {
        "tags": [
          "User"
        ],
        "summary": "Edits a user's details",
        "description": "This endpoint updates a user's details",
        "operationId": "userEdit",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "description": "Token Authentication",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "updated user object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/userUpdate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updates a user's details",
            "schema": {
              "$ref": "#/definitions/userUpdateRes"
            }
          },
          "400": {
            "description": "Invalid email address",
            "schema": {
              "$ref": "#/definitions/wrongEmail"
            }
          },
          "401": {
            "description": "Token issue",
            "schema": {
              "$ref": "#/definitions/invalidToken"
            }
          },
          "406": {
            "description": "Invalid Input",
            "schema": {
              "$ref": "#/definitions/invalidInput"
            }
          }
        }
      }
    },
    "/auth/logout": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Logs out a user",
        "description": "This endpoint logs out a user",
        "operationId": "userLogOut",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Updates a user's details",
            "schema": {
              "$ref": "#/definitions/logOut"
            }
          }
        }
      }
    },
    "/businesses": {
      "post": {
        "tags": [
          "Business"
        ],
        "summary": "Add a business",
        "description": "Add business to database.",
        "operationId": "createBusiness",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "description": "Token Authentication",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "body",
            "description": "Created business object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/createBusiness"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Business successfully created",
            "schema": {
              "$ref": "#/definitions/createBusinessRes"
            }
          },
          "401": {
            "description": "Token issue",
            "schema": {
              "$ref": "#/definitions/invalidToken"
            }
          },
          "406": {
            "description": "Invalid Input",
            "schema": {
              "$ref": "#/definitions/invalidInput"
            }
          },
          "409": {
            "description": "Business name exists",
            "schema": {
              "$ref": "#/definitions/existingBusiness"
            }
          }
        }
      },
      "get": {
        "tags": [
          "Business"
        ],
        "summary": "Get all businesses",
        "description": "Gets all businesses",
        "operationId": "getAllBusinesses",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Businesses found",
            "schema": {
              "$ref": "#/definitions/getAllBusinesses"
            }
          }
        }
      }
    },
    "/businesses/{businessId}": {
      "get": {
        "tags": [
          "Business"
        ],
        "summary": "Get business by businessId",
        "description": "Retrives a business",
        "operationId": "getBusiness",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "businessId",
            "in": "path",
            "description": "businessId of business to retrieve",
            "required": true,
            "type": "UUID"
          }
        ],
        "responses": {
          "200": {
            "description": "Business was successfully found",
            "schema": {
              "$ref": "#/definitions/getBusiness"
            }
          },
          "404": {
            "description": "Business not found!",
            "schema": {
              "$ref": "#/definitions/BusinessNotFound"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Business"
        ],
        "summary": "Update business profile",
        "description": "Updates business profile",
        "operationId": "updateBusiness",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "description": "Token Authentication",
            "required": false,
            "type": "string"
          },
          {
            "name": "businessId",
            "in": "path",
            "description": "business id to be updated",
            "required": true,
            "type": "UUID"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Updated business object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/updateBusiness"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Business successfully updated",
            "schema": {
              "$ref": "#/definitions/updatedBusiness"
            }
          },
          "401": {
            "description": "Token issue",
            "schema": {
              "$ref": "#/definitions/invalidToken"
            }
          },
          "404": {
            "description": "Business not found",
            "schema": {
              "$ref": "#/definitions/BusinessNotFound"
            }
          },
          "406": {
            "description": "Invalid Input",
            "schema": {
              "$ref": "#/definitions/invalidInput"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Business"
        ],
        "summary": "Remove a business",
        "description": "Deletes a business",
        "operationId": "deleteBusiness",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "description": "Token Authentication",
            "required": false,
            "type": "string"
          },
          {
            "name": "businessId",
            "in": "path",
            "description": "The business that would be deleted",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Business was successfully deleted",
            "schema": {
              "$ref": "#/definitions/deleteBusiness"
            }
          },
          "404": {
            "description": "Business not found",
            "schema": {
              "$ref": "#/definitions/BusinessNotFound"
            }
          }
        }
      }
    },
    "/businesses/?location=location": {
      "get": {
        "tags": [
          "Business"
        ],
        "summary": "Search businesses by location",
        "description": "This endpoint searches businesses by their location",
        "operationId": "getAllBusinesses",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "description": "Token Authentication",
            "required": false,
            "type": "string"
          },
          {
            "name": "location",
            "in": "query",
            "description": "Business category",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User successfully created",
            "schema": {
              "$ref": "#/definitions/businessSearch"
            }
          }
        }
      }
    },
    "/businesses/?category=category": {
      "get": {
        "tags": [
          "Business"
        ],
        "summary": "Search businesses by category",
        "description": "This endpoint searches businesses by their category",
        "operationId": "getAllBusinesses",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "description": "Token Authentication",
            "required": false,
            "type": "string"
          },
          {
            "name": "category",
            "in": "query",
            "description": "Business category",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Businesses found",
            "schema": {
              "$ref": "#/definitions/businessSearch"
            }
          }
        }
      }
    },
    "/businesses/?location=location&category=category": {
      "get": {
        "tags": [
          "Business"
        ],
        "summary": "Search businesses by location and category",
        "description": "This endpoint searches businesses by their location and category",
        "operationId": "getAllBusinesses",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "description": "Token Authentication",
            "required": false,
            "type": "string"
          },
          {
            "name": "location",
            "in": "query",
            "description": "Business category",
            "required": true,
            "type": "string"
          },
          {
            "name": "category",
            "in": "query",
            "description": "Business category",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Businesses found",
            "schema": {
              "$ref": "#/definitions/businessSearch"
            }
          }
        }
      }
    },
    "/businesses/{businessId}/reviews": {
      "post": {
        "tags": [
          "Review"
        ],
        "summary": "Add a review for a business",
        "description": "Add a review record for a business.",
        "operationId": "addReview",
        "consumes": [
          "application/x-www-form-urlencoded",
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "description": "Token Authentication",
            "required": false,
            "type": "string"
          },
          {
            "name": "businessId",
            "in": "path",
            "description": "business to add review to",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "201": {
            "description": "Review was successfully added",
            "schema": {
              "$ref": "#/definitions/addReview"
            }
          },
          "404": {
            "description": "Business was not found",
            "schema": {
              "$ref": "#/definitions/BusinessNotFound"
            }
          },
          "406": {
            "description": "Invalid Input",
            "schema": {
              "$ref": "#/definitions/invalidInput"
            }
          }
        }
      },
      "get": {
        "tags": [
          "Review"
        ],
        "summary": "Get all reviews for a business",
        "description": "Get all reviews for a business",
        "operationId": "getBusinessReview",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "description": "Token Authentication",
            "required": false,
            "type": "string"
          },
          {
            "name": "businessId",
            "in": "path",
            "description": "Business to be reviewed",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Reviews found",
            "schema": {
              "$ref": "#/definitions/getBusinessReview"
            }
          },
          "404": {
            "description": "Business was not found",
            "schema": {
              "$ref": "#/definitions/BusinessNotFound"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "firstname": {
          "type": "string",
          "example": "Tobiloba"
        },
        "lastname": {
          "type": "string",
          "example": "Emmanuel"
        },
        "email": {
          "type": "string",
          "example": "tobiemm@yahoo.com"
        }
      }
    },
    "userSignup": {
      "type": "object",
      "required": [
        "firstname",
        "lastname",
        "email",
        "password",
        "confirmPassword"
      ],
      "properties": {
        "firstname": {
          "type": "string",
          "example": "Tobiloba"
        },
        "lastname": {
          "type": "string",
          "example": "Emmanuel"
        },
        "email": {
          "type": "string",
          "example": "tobiemm@yahoo.com"
        },
        "password": {
          "type": "string",
          "example": "Tobiword"
        },
        "confirmPassword": {
          "type": "string",
          "example": "Tobiword"
        }
      }
    },
    "signupResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "User successfully created"
        },
        "user": {
          "type": "object",
          "$ref": "#/definitions/User"
        }
      }
    },
    "wrongEmail": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "Please enter a valid email address"
        }
      }
    },
    "invalidInput": {
      "type": "object",
      "properties": {
        "error": {
          "type": "array",
          "example": ["Error 1", "Error 2", "Error 3"]
        }
      }
    },
    "existingEmail": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "This email already has an account"
        }
      }
    },
    "userLogin": {
      "type": "object",
      "required": [
        "email",
        "password"
      ],
      "properties": {
        "email": {
          "type": "string",
          "example": "tobiemm@yahoo.com"
        },
        "password": {
          "type": "string",
          "example": "Tobiword"
        }
      }
    },
    "login": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "Welcome Tobiloba Emmanuel"
        }
      }
    },
    "wrongPassword": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "Wrong password"
        }
      }
    },
    "UserNotFound": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "Email not found"
        }
      }
    },
    "Business": {
      "type": "object",
      "properties": {
        "businessName": {
          "type": "string",
          "example": "Coffee Shop"
        },
        "businessImage": {
          "type": "string",
          "example": "https://images.unsplash.com/photo-1513866518382-c6e2c749de1a?ixlib=rb-0.3.5&ixid=eyJhcHBfaWQiOjEyMDd9&s=77739345ea1b858247d0d2a23b3bb328&auto=format&fit=crop&w=711&q=80"
        },
        "category": {
          "type": "string",
          "example": "Resturant"
        },
        "address": {
          "type": "string",
          "example": "10, Folorunsho Street"
        },
        "city": {
          "type": "string",
          "example": "Egbeda"
        },
        "state": {
          "type": "string",
          "example": "Lagos"
        },
        "phoneNumber": {
          "type": "string",
          "example": "07037439325"
        },
        "postalAddress": {
          "type": "string",
          "example": "P.O Box 115, Ikeja, Lagos"
        },
        "workHours": {
          "type": "string",
          "example": "8am to 4pm"
        },
        "about": {
          "type": "string",
          "example": "Lorem ipsum dolor sit amet, consectetur adipisicing elit"
        },
        "reviews": {
          "type": "array",
          "example": [1, 2, 3]
        },
        "createdAt": {
          "type": "string",
          "example": "3/10/2018, 10:40:17 PM"
        },
        "updateAt": {
          "type": "string",
          "example": "3/10/2018, 10:40:17 PM"
        }
      }
    },
    "getAllBusinesses": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "All Businesses"
        },
        "business": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Business"
          }
        }
      }
    },
    "createBusiness": {
      "type": "object",
      "required": [
        "businessName",
        "category",
        "address",
        "city",
        "state",
        "phoneNumber",
        "about"
      ],
      "properties": {
        "businessName": {
          "type": "string",
          "example": "Coffee Shop"
        },
        "businessImage": {
          "type": "string",
          "example": "https://images.unsplash.com/photo-1513866518382-c6e2c749de1a?ixlib=rb-0.3.5&ixid=eyJhcHBfaWQiOjEyMDd9&s=77739345ea1b858247d0d2a23b3bb328&auto=format&fit=crop&w=711&q=80"
        },
        "category": {
          "type": "string",
          "example": "Resturant"
        },
        "address": {
          "type": "string",
          "example": "10, Folorunsho Street"
        },
        "city": {
          "type": "string",
          "example": "Egbeda"
        },
        "state": {
          "type": "string",
          "example": "Lagos"
        },
        "phoneNumber": {
          "type": "string",
          "example": "07037439325"
        },
        "postalAddress": {
          "type": "string",
          "example": "P.O Box 115, Ikeja, Lagos"
        },
        "workHours": {
          "type": "string",
          "example": "8am to 4pm"
        },
        "about": {
          "type": "string",
          "example": "Lorem ipsum dolor sit amet, consectetur adipisicing elit"
        }
      }
    },
    "createBusinessRes": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example":"Business successfully created"
        },
        "business": {
          "type": "object",
          "$ref": "#/definitions/Business"
        }
      }
    },
    "BusinessNotFound": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "Business Not Found!"
        }
      }
    },
    "existingBusiness": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "Coffee Shop already exists as a business name"
        }
      }
    },
    "getBusiness": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "Business was successfully found"
        },
        "business": {
          "type": "object",
          "$ref": "#/definitions/Business"
        }
      }
    },
    "updateBusiness": {
      "type": "object",
      "properties": {
        "address": {
          "type": "string",
          "example": "10, Folorunsho Street"
        },
        "city": {
          "type": "string",
          "example": "Egbeda"
        },
        "about": {
          "type": "string",
          "example": "Lorem ipsum dolor sit amet, consectetur adipisicing elit"
        }
      }
    },
    "updatedBusiness": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "Business successfully updated"
        },
        "business": {
          "type": "object",
          "$ref": "#/definitions/Business"
        }
      }
    },
    "deleteBusiness": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "Business was successfully deleted"
        },
        "business": {
          "type": "object",
          "$ref": "#/definitions/Business"
        }
      }
    },
    "businessSearch": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "Businesses found"
        },
        "business": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Business"
          }
        }
      }
    },
    "Review": {
      "type": "object",
      "properties": {
        "review": {
          "type": "string",
          "example": "T'was nice doing business with you guys"
        },
        "reviewId": {
          "type": "integer",
          "example": 1
        },
        "businessId":{
          "type": "integer",
          "example": 1
        },
        "createdAt": {
          "type": "string",
          "example": "3/10/2018, 10:40:17 PM"
        },
        "updateAt": {
          "type": "string",
          "example": "3/10/2018, 10:40:17 PM"
        }
      }
    },
    "addReview": {
      "type": "object",
      "required": [
        "review"
      ],
      "properties": {
        "review": {
          "type": "string",
          "example": "T'was nice doing business with you guys"
        }
      }
    },
    "addReviewRes": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "Review was successfully added"
        },
        "review": {
          "type": "object",
          "$ref": "#/definitions/Review"
        }
      }
    },
    "getBusinessReview": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "Reviews found"
        },
        "reviews": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Review"
          }
        }
      }
    }
  },
  "externalDocs": {
    "description": "",
    "url": ""
  }
}